requirements.txt
# ═══════════════════════════════════════════
#  Attendance AI — Python Dependencies
# ═══════════════════════════════════════════

# ── Web Framework ──
fastapi==0.131.0
uvicorn[standard]==0.41.0
python-multipart==0.0.22
pydantic-settings==2.13.1

# ── Database ──
motor==3.7.1
pymongo==4.16.0

# ── Face Recognition (GPU) ──
facenet-pytorch==2.6.0
faiss-cpu==1.13.2

# ── Deep Learning (CUDA) ──
# Install PyTorch with CUDA from: https://pytorch.org/get-started/locally/
# Example: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu128
torch>=2.0.0
torchvision>=0.15.0

# ── Data Processing ──
numpy>=2.0.0
Pillow>=10.0.0
pandas>=2.0.0
